Remove abstract messaging and flow builder modules
The modules `abstract.py` and `flow_builder.py` were removed from the `common/messaging` package. This eliminates unused or unnecessary implementations related to messaging abstractions and flow building.
