+++ common/messaging/platform_router.py
@@ -5,7 +5,7 @@ import importlib
 import inspect
 from typing import Dict, Any, Optional, Union, Callable, List, Tuple, Type
 
-from common.messaging.interface import MessagingInterface
+from common.messaging.unified_interface import MessagingInterface
 from common.messaging.unified_platform_utils import resolve_user_id
 
 logger = logging.getLogger(__name__)
