+++ services/whatsapp_service/app/flow_integration.py
@@ -4,8 +4,8 @@ import logging
 import asyncio
 from twilio.twiml.messaging_response import MessagingResponse
 from .bot import sanitize_phone_number, get_user_state, update_user_state
-from common.messaging.flow_builder import flow_library
-from common.messaging.flow_integration_helper import (
+from common.messaging.unified_flow import flow_library
+from common.messaging.unified_flow import (
     check_and_process_flow,
     process_flow_action,
     show_available_flows
